requirements.txt
argcomplete==3.3.0
asgiref==3.7.2
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distlib==0.3.8
django-cors-headers==4.4.0
django-countries==7.6.1
django-js-asset==2.2.0
djangorestframework==3.15.2
filelock==3.13.4
idna==3.7
jmespath==1.0.1
packaging==24.0
pathspec==0.10.1
platformdirs==4.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
userpath==1.9.2
wcwidth==0.1.9
